requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Machine Learning & Vector Search
xgboost>=2.0.0
faiss-cpu>=1.7.4

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
joblib>=1.3.0